refactor: extract debug metadata builder from llvm-generator (~180 lines)
